Move modifiers module to top level
The modifiers module has no direct relation to the serving over HTTP.
